Split blackhole.c into several separate files
